+++ tenann/searcher/faiss_ivf_pq_ann_searcher.h
@@ -34,10 +34,10 @@ class FaissIvfPqAnnSearcher : public AnnSearcher {
   T_FORBID_COPY_AND_ASSIGN(FaissIvfPqAnnSearcher);
 
   /// ANN搜索接口，只返回k近邻的id
-  void AnnSearch(PrimitiveSeqView query_vector, int k, int64_t* result_id,
+  void AnnSearch(PrimitiveSeqView query_vector, int64_t k, int64_t* result_id,
                  const IdFilter* id_filter = nullptr) override;
 
-  void AnnSearch(PrimitiveSeqView query_vector, int k, int64_t* result_ids,
+  void AnnSearch(PrimitiveSeqView query_vector, int64_t k, int64_t* result_ids,
                  uint8_t* result_distances, const IdFilter* id_filter = nullptr) override;
 
   void RangeSearch(PrimitiveSeqView query_vector, float range, int64_t limit,
